pyproject.toml
[project]
name = "kishin-api"
version = "0.1.0"
description = ""
authors = [
    {name = "Maxime Puys", email = "drakes00@gmail.com"}
]
readme = "README.md"
requires-python = ">=3.13,<4.0.0"
dependencies = [
    "fastapi>=0.128.6,<0.129.0",
    "gpx>=2026.1.0,<2027.0.0",
    "alpha-shapes>=1.1.1,<2.0.0",
    "matplotlib>=3.10.8,<4.0.0",
    "pyqt6>=6.10.2,<7.0.0",
    "h3>=4.4.2,<5.0.0",
    "requests>=2.32.5,<3.0.0",
    "geopandas>=1.1.2,<2.0.0",
    "uvicorn>=0.41.0,<0.42.0",
    "pyjwt[crypto]>=2.11.0,<3.0.0",
    "pwdlib[argon2]>=0.3.0,<0.4.0",
    "sqlalchemy>=2.0.46,<3.0.0",
    "python-multipart>=0.0.22,<0.0.23",
    "pydantic-settings (>=2.13.1,<3.0.0)",
    "pytest-cov (>=7.0.0,<8.0.0)",
]

[tool.poetry]
packages = [{include = "kishin_trails"}]

[tool.poetry.group.dev.dependencies]
ipython = ">=9.10.0,<10.0.0"
httpx = ">=0.28.1,<0.29.0"
pytest = ">=9.0.2,<10.0.0"
pytest_asyncio = ">=1.3.0,<2.0.0"
pylint = "^4.0.5"

[build-system]
requires = ["poetry-core>=2.0.0,<3.0.0"]
build-backend = "poetry.core.masonry.api"
